+++ tests/internals/idle_replicas/idle_replicas_test.go
@@ -1,7 +1,7 @@
 //go:build e2e
 // +build e2e
 
-package azure_queue_restore_original_test
+package idle_replicas_test
 
 import (
 	"context"
@@ -9,6 +9,7 @@ import (
 	"fmt"
 	"os"
 	"testing"
+	"time"
 
 	"github.com/Azure/azure-storage-queue-go/azqueue"
 	"github.com/joho/godotenv"
@@ -26,7 +27,7 @@ import (
 var _ = godotenv.Load("../../.env")
 
 const (
-	testName = "azure-queue-restore-test"
+	testName = "idle-replicas-test"
 )
 
 var (
@@ -68,7 +69,7 @@ metadata:
   labels:
     app: {{.DeploymentName}}
 spec:
-  replicas: 2
+  replicas: 0
   selector:
     matchLabels:
       app: {{.DeploymentName}}
@@ -98,12 +99,11 @@ metadata:
   name: {{.ScaledObjectName}}
   namespace: {{.TestNamespace}}
 spec:
-  advanced:
-    restoreToOriginalReplicaCount: true
   scaleTargetRef:
     name: {{.DeploymentName}}
   pollingInterval: 5
-  minReplicaCount: 0
+  idleReplicaCount: 0
+  minReplicaCount: 2
   maxReplicaCount: 4
   cooldownPeriod: 10
   triggers:
@@ -117,9 +117,9 @@ spec:
 func TestScaler(t *testing.T) {
 	// setup
 	t.Log("--- setting up ---")
-	require.NotEmpty(t, connectionString, "AZURE_STORAGE_CONNECTION_STRING env variable is required for service bus tests")
+	require.NotEmpty(t, connectionString, "AZURE_STORAGE_CONNECTION_STRING env variable is required for idle replicas test")
 
-	queueURL := createQueue(t)
+	queueURL, messageURL := createQueue(t)
 
 	// Create kubernetes resources
 	kc := GetKubernetesClient(t)
@@ -127,19 +127,22 @@ func TestScaler(t *testing.T) {
 
 	CreateKubernetesResources(t, kc, testNamespace, data, templates)
 
-	assert.True(t, WaitForDeploymentReplicaCount(t, kc, deploymentName, testNamespace, 2, 60, 1),
-		"replica count should be 2 after a minute")
+	// scaling to idle replica count
+	assert.True(t, WaitForDeploymentReplicaCount(t, kc, deploymentName, testNamespace, 0, 60, 1),
+		"replica count should be 0 after a minute")
 
 	// test scaling
-	testScale(t, kc, data)
-	testRestore(t, kc, data)
+	// till min replica count
+	testScaleUp(t, kc, messageURL)
+	// back to idle replica count
+	testScaleDown(t, kc, messageURL)
 
 	// cleanup
 	DeleteKubernetesResources(t, kc, testNamespace, data, templates)
 	cleanupQueue(t, queueURL)
 }
 
-func createQueue(t *testing.T) azqueue.QueueURL {
+func createQueue(t *testing.T) (azqueue.QueueURL, azqueue.MessagesURL) {
 	// Create Queue
 	httpClient := kedautil.CreateHTTPClient(DefaultHTTPTimeOut, false)
 	credential, endpoint, err := azure.ParseAzureStorageQueueConnection(
@@ -154,7 +157,9 @@ func createQueue(t *testing.T) azqueue.QueueURL {
 	_, err = queueURL.Create(context.Background(), azqueue.Metadata{})
 	assert.NoErrorf(t, err, "cannot create storage queue - %s", err)
 
-	return queueURL
+	messageURL := queueURL.NewMessagesURL()
+
+	return queueURL, messageURL
 }
 
 func getTemplateData() (templateData, templateValues) {
@@ -168,24 +173,30 @@ func getTemplateData() (templateData, templateValues) {
 			ScaledObjectName: scaledObjectName,
 			QueueName:        queueName,
 		}, templateValues{
-			"secretTemplate":     secretTemplate,
-			"deploymentTemplate": deploymentTemplate}
+			"secretTemplate":       secretTemplate,
+			"deploymentTemplate":   deploymentTemplate,
+			"scaledObjectTemplate": scaledObjectTemplate}
 }
 
-func testScale(t *testing.T, kc *kubernetes.Clientset, data templateData) {
-	t.Log("--- testing scaling ---")
-	KubectlApplyWithTemplate(t, data, "scaledObjectTemplate", scaledObjectTemplate)
-
-	assert.True(t, WaitForDeploymentReplicaCount(t, kc, deploymentName, testNamespace, 0, 60, 1),
-		"replica count should be 0 after a minute")
-}
-
-func testRestore(t *testing.T, kc *kubernetes.Clientset, data templateData) {
-	t.Log("--- testing restore ---")
-	KubectlDeleteWithTemplate(t, data, "scaledObjectTemplate", scaledObjectTemplate)
+func testScaleUp(t *testing.T, kc *kubernetes.Clientset, messageURL azqueue.MessagesURL) {
+	t.Log("--- testing scale up ---")
+	for i := 0; i < 5; i++ {
+		msg := fmt.Sprintf("Message - %d", i)
+		_, err := messageURL.Enqueue(context.Background(), msg, 0*time.Second, time.Hour)
+		assert.NoErrorf(t, err, "cannot enqueue message - %s", err)
+	}
 
 	assert.True(t, WaitForDeploymentReplicaCount(t, kc, deploymentName, testNamespace, 2, 60, 1),
 		"replica count should be 2 after a minute")
+}
+
+func testScaleDown(t *testing.T, kc *kubernetes.Clientset, messageURL azqueue.MessagesURL) {
+	t.Log("--- testing scale down ---")
+	_, err := messageURL.Clear(context.Background())
+	assert.NoErrorf(t, err, "cannot clear queue - %s", err)
+
+	assert.True(t, WaitForDeploymentReplicaCount(t, kc, deploymentName, testNamespace, 0, 60, 1),
+		"replica count should be 0 after a minute")
 }
 
 func cleanupQueue(t *testing.T, queueURL azqueue.QueueURL) {
